requirements.txt
asgiref==3.4.1
bleach==4.1.0
cached-property==1.5.2
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.6
django-admin-shell==1.0.0
django-attachments==1.9.1
django-debug-toolbar==3.2.2
django-environ==0.8.1
django-extensions==3.1.3
django-filter==21.1
django-import-export==2.6.1
django-model-utils==4.2.0
django-modeltranslation==0.17.3
django-phonenumber-field==5.2.0
django-seed==0.3.1
django-simple-history==3.0.0
django-sluggable==0.7.0
django-sql-dashboard==1.0.1
django-sql-explorer==2.3
django-url-filter==0.3.15
enum-compat==0.0.3
et-xmlfile==1.1.0
Faker==9.8.2
importlib-metadata==4.8.2
Markdown==3.3.6
MarkupPy==1.14
mock==4.0.3
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
phonenumbers==8.12.36
psycopg2==2.9.1
psycopg2-binary==2.9.1
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
text-unidecode==1.3
toposort==1.7
typing-extensions==3.10.0.2
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.6.0
